Rename legacy types to legacy to eventually remove them
