+++ client/src/components/DeclarationOfDataConsent/DeclarationOfDataConsent.tsx
